Add files via upload
This file can be divided into 4 main parts:-

In part no. 1, the code is using Selenium WebDriver to scrape product names and reviews from the webpage of Daraz. It navigates to the URL of the webpage, iterates over the products on the page, and extracts the product name and review text. If a review is not available, it records ‘Not Available’. The code runs this process for three pages, navigating to the next page after each iteration. The scraped data is stored in a pandas DataFrame. <br><br>

In part no. 2, the code is performing sentiment analysis on the reviews scraped in Q1. It uses the Sentiment Intensity Analyzer (sid) from the nltk library to calculate polarity scores for each review. The compound score is extracted from these polarity scores and used to classify each review as positive (‘pos’) or negative (‘neg’). <br><br>

In Part no. 3, the code is scraping product details from the Daraz website for a specified product (in this case, ‘mobile covers’). It navigates to the search results page for the product, then iterates over the products on the page, extracting details such as the product name, price, rating, number of reviews, availability on Daraz Mall, and country of origin. The scraped data is stored in again stored in a pandas DataFrame.<br><br>

In the last part, i.e. Part no. 4, code is scraping detailed product information from the Daraz website for a specified product (in this case, ‘Mobile Covers’). It navigates to the search results page for the product, then iterates over the products on the page, extracting details such as the product name, whether it’s available on Daraz Mall or from a private seller, average price, number of listings, average unit selling, seller ratings, country of origin, and number of followers. <br>

The code also navigates to each product’s individual page to gather more detailed information. After gathering the information for a product, it navigates back to the search results page and undoes the selection of the current product before moving on to the next one. The scraped data is again stored in a pandas DataFrame.<br>